fix(jsonl): remove needless array nesting
Fixes #261. Before, the code was writing an empty array (`[]`) if the
`nucleicMd5` had already been seen. Now that the array has been removed,
that's no longer possible, so I've added an `else` block that writes a
`null` value instead. With that being said, I actually have a hunch that
generating an empty array was always buggy behaviour, so I'll (attempt
to) address that in my upcoming commit.